fix: Remove unused dependencies and files
