+++ packages/altfire_tracker/example/lib/record_error_interceptor.dart
@@ -1,18 +1,5 @@
 import 'package:altfire_tracker/altfire_tracker.dart';
 import 'package:dio/dio.dart';
-import 'package:riverpod_annotation/riverpod_annotation.dart';
-
-part 'interceptor_provider.g.dart';
-
-@Riverpod(keepAlive: true)
-LogInterceptor logInterceptor(LogInterceptorRef ref) => LogInterceptor(
-      requestBody: true,
-      responseBody: true,
-    );
-
-@Riverpod(keepAlive: true)
-RecordErrorInterceptor recordErrorInterceptor(RecordErrorInterceptorRef ref) =>
-    RecordErrorInterceptor();
 
 /// Interception class for sending error logs
 /// when an error occurs during API communication
