+++ packages/altfire_tracker/example/pubspec.yaml
@@ -13,16 +13,12 @@ dependencies:
   firebase_core: ^2.21.0
   flutter:
     sdk: flutter
-  hooks_riverpod: ^2.4.9
-  riverpod_annotation: ^2.3.3
 
 dev_dependencies:
   altive_lints: ^1.9.0
-  build_runner: ^2.4.8
   flutter_test:
     sdk: flutter
   mockito: ^5.4.4
-  riverpod_generator: ^2.3.9
 
 flutter:
   uses-material-design: true
